% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_xena.R
\name{import_xena_cohort}
\alias{import_xena_cohort}
\title{极速构建 bulkKnk 标准化分析队列 (适配 UCSC Xena)}
\usage{
import_xena_cohort(expr_file, surv_file, convert_id = TRUE, species = "human")
}
\arguments{
\item{expr_file}{字符串。UCSC Xena 表达矩阵文件路径 (如 .tsv.gz)。}

\item{surv_file}{字符串。UCSC Xena 生存数据文件路径 (如 survival.tsv)。}

\item{convert_id}{逻辑值。是否自动将 Ensembl ID 转换为 SYMBOL，默认 TRUE。}

\item{species}{字符串。物种，默认 "human"。用于 ID 转换。}
}
\value{
一个 bulkKnk_cohort (S3) 对象，包含:
\itemize{
\item \code{expr_matrix}: 清洗完毕的表达矩阵 \link{行=样本, 列=基因}
\item \code{clinical_data}: 完美对齐的生存数据框 (包含 OS_time, OS_status)
\item \code{pheno_vector}: 默认生成的全 Tumor 表型向量
}
}
\description{
本函数旨在提供“一键式”的数据加载体验。它会极速读取 UCSC Xena 下载的
表达矩阵和生存数据，自动完成 Ensembl -> SYMBOL 的转换、去除死寂基因、
矩阵转置，并将生存数据与表达矩阵进行绝对严格的样本对齐。
}
