% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_predictor.R
\name{plot_cox_forest}
\alias{plot_cox_forest}
\title{绘制预后签名 Cox 回归森林图}
\usage{
plot_cox_forest(result, top_n = 15)
}
\arguments{
\item{result}{predict_ko_phenotype() 的返回值}

\item{top_n}{展示前 N 个最显著的签名，默认 15}
}
\value{
ggplot2 森林图
}
\description{
绘制预后签名 Cox 回归森林图
}
